move hubs to api project
