--- PRJ/HopeLine.Service/Hubs/ScheduleHub.cs
@@ -1,8 +1,0 @@
-﻿using Microsoft.AspNetCore.SignalR;
-
-namespace HopeLine.Service.Hubs
-{
-    public class ScheduleHub : Hub
-    {
-    }
-}
